fix filters behavior
